refactor: switch transformers executor for hub version
